Changing "Concatenation" to a module.
First brick of a change in the long term.
It will help to monitor the changes step by step rather than fighting with the existing python file.

Add a "utils" with ugly "isinstance" function to avoid circular imports inside the function.

Still less ugly than a 5000 lines with intricated class.